requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
apscheduler==3.10.4
pyotp==2.9.0
cryptography==42.0.5
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.9
websockets==12.0
